renamed file from products.ts to product.ts
