+++ backend/src/app.ts
@@ -2,7 +2,7 @@ import express from "express";
 import cors from "cors";
 import mongoose from 'mongoose';
 import path from 'path';
-import productsRouter from './routes/product';
+import productRouter from './routes/product';
 
 const { PORT = 3000, DB_ADDRESS = 'mongodb://127.0.0.1:27017/weblarek', UPLOAD_PATH, UPLOAD_PATH_TEMP } = process.env;
 
@@ -18,7 +18,7 @@ app.use(express.urlencoded({ extended: true }));
 
 app.use(express.static(path.join(__dirname, 'public')));
 
-app.use(productsRouter);
+app.use(productRouter);
 
 app.listen(PORT, () => {
   console.log("listening on port 3000");
